Fix Mouse::Hide for Win32.
The ShowCursor function works quiet tricky.
There is a counter, which is initialized to 0.
if counter >= 0 the cursor is visible.
if counter <= -1 the cursor is hidden.
ShowCursor(true) increments and ShowCursor(false) decrements the counter.
When entering grab, the counter will decrement by 2,
When leaving grab the counter will increment by 2.
So cursor will only be visible in grab(false) mode.